Support for Boolean data arrays in DiscreteProperty
Fix #327
Renamed integer getters and setters of ValuesProperty to honor C++11 names.
Replace some unsigned int with uint64_t

+++ src/eml2/TimeSeries.cpp
@@ -31,7 +31,7 @@ void TimeSeries::pushBackTimestamp(time_t timestamp)
 	pushBackTimestamp(tmConversion);
 }
 
-time_t TimeSeries::getTimestamp(unsigned int index) const
+time_t TimeSeries::getTimestamp(uint64_t index) const
 {
 	tm temp = getTimestampAsTimeStructure(index);
 	return timeTools::timegm(temp);
